Refactoring Catalog component. Create a CatalogLayerMenu.vue to handle single component test
